+++ CHANGELOG.md
@@ -10,6 +10,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - **cf-expandables:** [PATCH] Update to ES6 syntax.
 - **cf-tables:** [PATCH] Update to ES6 syntax.
 - **cf-atomic-component:** [PATCH] Reference cf-atomic-component locally.
+- **capital-framework:** [PATCH] Use uglifyjs-webpack-plugin.
 
 ### Removed
 - **cf-expandables:** [PATCH] Remove on-ready check.